% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_likert_table.R
\name{cross.likert.table}
\alias{cross.likert.table}
\title{Cross likert table}
\usage{
cross.likert.table(data, data.character, factor.list, weights, exact = FALSE,
  output.format = 2, na.rm = FALSE, na.string = "Trudno powiedzieć")
}
\arguments{
\item{data}{dane (pełen zakres)}

\item{data.character}{wybrane kolumny (nazwy kolumn lub początek nazw)}

\item{factor.list}{przecięcia (uwaga! zmienne muszą mieć atrybut "labels"!)}

\item{weights}{wagi}

\item{exact}{jeżeli w data.character podano jedną kolumnę, to exact = TRUE}

\item{output.format}{1 - lista, 2 - data frame}

\item{na.rm}{czy usunąć dane opisane jako na.string}

\item{na.string}{dane do usunięcia (domyślnie "Trudno powiedzieć")}
}
\description{
Tworzy dane częstości występowania odpowiedzi w pytaniach ze skali Likerta
}
