% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_table.R
\name{cross.table}
\alias{cross.table}
\title{Cross table}
\usage{
cross.table(data, data.character = NULL, factor.list, weights = NULL)
}
\arguments{
\item{data}{dane w formie data frame}

\item{data.character}{lista kolumn do analizy (jeżeli zmienna data zawiera cały zbiór danych)}

\item{factor.list}{zmienne do przecięcia danych}

\item{weights}{wagi}
}
\description{
Cross table
}
